deleted seperate ratings page
