+++ code-assets/backend/data/json_rating/COMP_rating.json
@@ -26,5 +26,20 @@
             "mandatory": "Yes",
             "review": "Worst Professor!"
         }
+    ],
+    "Cecile Bendavid": [
+        {
+            "professor_first_name": "Cecile",
+            "professor_last_name": "Bendavid",
+            "subject": "COMP",
+            "catalog_number": "Comp 110",
+            "star_rating": 4,
+            "grade": "B+",
+            "difficulty": 4,
+            "retake_professor": "Yes",
+            "require_textbooks": "Yes",
+            "mandatory": "Yes",
+            "review": "Good Professor."
+        }
     ]
 }
