+++ code-assets/frontend/csun-dash/src/App.js
@@ -1,12 +1,11 @@
 import React from 'react';
 import HomePage from './pages/HomePage'
 import PlannerPage from './pages/PlannerPage';
-import RatingsPage from './pages/RatingsPage';
 import ProfessorSearchPage from './pages/ProfessorSearchPage';
 import CS from './elements/majors/CS';
 import SignInPage from './pages/SignInPage'
 import {BrowserRouter as Router, Routes, Route} from "react-router-dom"
-import RatingsPageTwo from './pages/RatingsPageTwo';
+import RatingsPage from './pages/RatingsPage';
 
 function App() {
   return (
@@ -15,8 +14,7 @@ function App() {
           <Routes>
             <Route path='/' element={<HomePage/>}/>
             <Route path='/planner' element={<PlannerPage/>}/>
-            <Route path='/ratings' element={<RatingsPage/>}/>
-            <Route path='/ratings-two/:subject/:first_name/:last_name' element={<RatingsPageTwo/>} />
+            <Route path='/ratings/:subject/:first_name/:last_name' element={<RatingsPage/>} />
             <Route path='/professor-search' element={<ProfessorSearchPage/>}/>
             <Route path='/cs' element={<CS/>}/>
             <Route path='/signIn' element={<SignInPage/>}/>
