+++ code-assets/frontend/csun-dash/src/elements/Professor Search/ProfessorsSearchBody.js
@@ -20,7 +20,7 @@ function ProfessorsSearchPage({allProfessors}){
     }, [allProfessors])
 
     return(
-        <div style={{margin:"100px 100px 0px 100px", paddingBottom: "50px"}}>
+        <div style={{margin:"100px 100px 0px 100px", paddingBottom: "50px", boxShadow: "rgba(0, 0, 0, 0.25) 0px 54px 55px, rgba(0, 0, 0, 0.12) 0px -12px 30px, rgba(0, 0, 0, 0.12) 0px 4px 6px, rgba(0, 0, 0, 0.17) 0px 12px 13px, rgba(0, 0, 0, 0.09) 0px -3px 5px"}}>
             <TableContainer component={Paper}>
                 <Table>
                     <TableHead style={{backgroundColor: "#E31C25"}}>
